requirements.txt
python-telegram-bot==20.8
telethon==1.29.2
